# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sponsor": {
            "get": {
                "tags": [
                    "Sponsors"
                ],
                "summary": "Show all sponsors",
                "operationId": "807ae4de1f50e69c31f7628071c407cd",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sponsors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sponsor/create": {
            "post": {
                "tags": [
                    "Sponsors"
                ],
                "summary": "create a new sponsor",
                "operationId": "8f2fd5ec3ed61e8be2bacfcd10e61874",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "title of sponsors",
                                        "type": "string"
                                    },
                                    "desc": {
                                        "description": "sponsor description",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "sponsor image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "expired_at": {
                                        "description": "what is the sponsor expired at",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "comment": {
                                        "description": "to access image use https://mahllola.online/public/image  example : https://mahllola.online/public/storage/services_folder/ttyVNuauz67kqXX40jyewMwh3DWpdFjjyJ0pIiPd.jpg",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "add new sponsor successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sponsor/{id}": {
            "get": {
                "tags": [
                    "Sponsors"
                ],
                "summary": "Get sponsor by id",
                "operationId": "1ba9a15657ee63f87e8ccac35bbe6391",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "sponsor id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sponsors"
                ],
                "summary": "Delete an sponsor",
                "description": "Delete sponsor by ID",
                "operationId": "67091bdbfdcf3ca08911c5cd62dc6a94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sponsor to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sponsor deleted successfully"
                    },
                    "404": {
                        "description": "sponsor not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}